+++ multimessenger/__init__.py
@@ -1,0 +1,5 @@
+from .supernova import *
+try:
+    from ._version import version as __version__
+except ImportError:
+    pass
