+++ multimessenger/supernova/_version.py
