+++ multimessenger/supernova/constants.py
