+++ multimessenger/supernova/libraries.py
