+++ requirements.txt
@@ -1,5 +1,5 @@
 
-astropy~=5.0.1
+astropy~=4.3.1
 matplotlib~=3.5.1
 seaborn~=0.11.2
 scipy~=1.8.0
