+++ setup.cfg
@@ -1,11 +1,7 @@
 [metadata]
-name = snx
+name = multimessenger
 keywords =
     setup
     distutils
 
-
-[files]
-packages =
-    snx
